Improved helper functions, Found that they can be added to records :-)
