+++ Epsilon.Canvas.Abstractions/Model/Outcome.cs
@@ -1,4 +1,5 @@
 ﻿using System.Text.Json.Serialization;
+using System.Text.RegularExpressions;
 
 namespace Epsilon.Canvas.Abstractions.Model;
 
@@ -6,4 +7,23 @@ public record Outcome(
     [property: JsonPropertyName("id")] int Id,
     [property: JsonPropertyName("title")] string Title,
     [property: JsonPropertyName("description")] string Description
-);
+)
+{
+    public string ShortDescription()
+    {
+        string description = RemoveHtml();
+        //Function gives only the short English description back of the outcome. 
+        var startPos = description.IndexOf(" EN ", StringComparison.Ordinal) + " EN ".Length;
+        var endPos = description.IndexOf(" NL ", StringComparison.Ordinal);
+
+        return description.Substring(startPos, endPos - startPos);
+    }
+
+    private string RemoveHtml()
+    {
+        var raw = Regex.Replace(Description, "<.*?>", " ");
+        var trimmed = Regex.Replace(raw, @"\s\s+", " ");
+
+        return trimmed;
+    }
+};
